Provide estimate of rate limit remaining
As requests are tracked per-VM, there will be variation in the number of
requests remaining based on which VM was hit. To make this a bit less
variable, calculate a rough estimate (to the nearest 10%) for how many
requests a user has left. This won't be totally accurate and may vary
slightly across VMs but will vary significantly less due to rounding.
This limit uses floor() so will drop to 0 before the rate limit is
actually reached but this is safer than telling a user they have some
limit remaining when they do not. The limit header should reflect the
global amount but the per-VM calculation for remaining amount/whether
the limit has been exceeded should use the per-VM limit

Co-Authored-By: Marc Misoch <marc.misoch@sap.com>