Offset reset interval based on user GUID
If we reset all rate limits at the same time, there is a risk that many
rate limited users will all retry as soon as the reset interval ends. To
avoid this, use the user GUID to create an offset between 0-1 *
reset_interval to distribute reset times roughly evenly across the reset
interval.